Change implementation of key creation
Rather than using a hash of the file contents (which will eventually
blow up as soon as we have a collision, which will cause us to have to
move to a different implementation for `resultMap`), we'll switch over
to using `string` (still aliased to `resKey`) as keys for our map. This
will allow us to easily swap in a more robust key creation
implementation later down the line as well.